Move the permission get/set out of the common code
Instead, introduce a new v9_common.h header that has the v9 permission
encoding and implement a Morello version that does not rely on the UPERMS
and HWPERMS fields. This also means we no longer need to define the
USES_V9_PERMISSION_ENCODING macro in each header.
