+++ cheri_compressed_cap_128.h
@@ -155,10 +155,10 @@ _CC_STATIC_ASSERT_SAME(CC128_MANTISSA_WIDTH, CC128_FIELD_EXP_ZERO_BOTTOM_SIZE);
 #define CC128_RESERVED_BITS CC128_FIELD_RESERVED_SIZE
 #define CC128_HAS_BASE_TOP_SPECIAL_CASES 0
 #define CC128_USES_V9_CORRECTION_FACTORS 1
-#define CC128_USES_V9_PERMISSION_ENCODING 1
 #define CC128_USES_LEN_MSB 0
 
 #include "cheri_compressed_cap_common.h"
+#include "cheri_compressed_cap_v9_common.h"
 
 // Sanity-check mask is the expected NULL encoding
 _CC_STATIC_ASSERT_SAME(CC128_MEM_XOR_MASK, UINT64_C(0x00001ffffc018004));
