+++ cheri_compressed_cap_128r.h
@@ -146,7 +146,6 @@ _CC_STATIC_ASSERT_SAME(CC128R_MANTISSA_WIDTH, CC128R_FIELD_EXP_ZERO_BOTTOM_SIZE)
 #define CC128R_RESERVED_FIELDS 2
 #define CC128R_RESERVED_BITS (CC128R_FIELD_RESERVED0_SIZE + CC128R_FIELD_RESERVED1_SIZE)
 #define CC128R_HAS_BASE_TOP_SPECIAL_CASES 1
-#define CC128R_USES_V9_PERMISSION_ENCODING 0
 #define CC128R_USES_V9_CORRECTION_FACTORS 0
 #define CC128R_USES_LEN_MSB 0
